requirements.txt
elastic-transport==8.12.0
elasticsearch==8.12.1
python-dotenv==1.0.1
azure-core==1.30.1
azure-identity==1.15.0
azure-keyvault-secrets==4.8.0